requirements.txt
pytest>=7.0.0
requests>=2.28.0
allure-pytest>=2.9.45
pyyaml>=6.0